requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.6
authlib==1.0.0rc1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
click==8.0.4; python_version >= '3.6'
cryptography==36.0.1; python_version >= '3.6'
flask==2.0.3; python_version >= '3.6'
fastapi==0.74.1
uvicorn==0.17.5
greenlet==2.0.0a1; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.3; python_version >= '3'
importlib-metadata==4.11.1; python_version < '3.9'
importlib-resources==5.4.0; python_version < '3.9'
itsdangerous==2.1.0; python_version >= '3.7'
jinja2==3.0.3; python_version >= '3.6'
kafka-python==2.0.2
mako==1.1.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown2==2.4.2
markupsafe==2.1.0; python_version >= '3.7'
mypy-extensions==0.4.3
origin-platform-utils==0.6.0
psycopg2==2.9.3; python_version >= '3.6'
pycparser==2.21
pycryptodome==3.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyjwt==2.3.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.6
python-rapidjson==1.6; python_version >= '3.6'
pytz==2021.3
rapidjson==1.0.0
requests==2.27.1
serpyco==1.3.5; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.31; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
typing-extensions==4.1.1; python_version >= '3.6'
typing-inspect==0.7.1
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.3; python_version >= '3.6'
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.7.0; python_version >= '3.7'
